pyproject.toml
[project]
name = "agent-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "mlflow>=3.7.0",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-ai>=1.37.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "scikit-learn>=1.8.0",
    "uvicorn[standard]>=0.32.0",
]
